requirements.txt
tensorflow-gpu
scipy
graphviz
pydot
matplotlib
numpy